2017-08-25 Roman (stable) - !!!!:SYS.Transports,Transports.*,DAQ.{AMRDevs,DCON,MMS,OPC_UA,SMH2Gi,Siemens,ICP_DAS,FT3},Protocol.{ModBus,HTTP,SelfSystem,UserProtocol}:     The generic accessing resource was moved to mutex and an argument "noRes" was removed from function messIO()     to prevent from unexpected outcomes from the requesting critical section.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2511 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/VCAEngine/session.cpp
@@ -557,13 +557,13 @@ TVariant Session::objFuncCall( const string &iid, vector<TVariant> &prms, const 
 	return 0;
     }
     // int reqTm( ) - Last request time
-    if(iid == "reqTm")		return reqTm();
+    if(iid == "reqTm")		return (int64_t)reqTm();
     // string reqUser( ) - Last request user
     if(iid == "reqUser")	return reqUser();
     // string reqLang( ) - Last request language
     if(iid == "reqLang")	return reqLang();
     // int userActTm( ) - Last user action time
-    if(iid == "userActTm")	return userActTm();
+    if(iid == "userActTm")	return (int64_t)userActTm();
 
     return TCntrNode::objFuncCall(iid,prms,cuser);
 }
